feat: nest "friends" under "connections" feature
